+++ client/src/features/my/api/useChangeNicknameMutation.ts
@@ -1,8 +1,8 @@
 import { useMutation } from '@tanstack/react-query';
-import { changeNickname } from '../api/changeNickname';
 import { queryClient } from '@/app/lib/queryClient';
 import { ChangeNicknameRequest } from '../types/changeNickname';
 import { useToast } from '@/shared/hooks';
+import { api } from '@/app/lib/api';
 
 export const useChangeNicknameMutation = () => {
   const { showSuccess, showError } = useToast();
@@ -17,4 +17,9 @@ export const useChangeNicknameMutation = () => {
       showError('닉네임 변경에 실패했습니다. 다시 시도해주세요.');
     },
   });
+};
+
+export const changeNickname = async ({ newNickname }: ChangeNicknameRequest) => {
+  const response = await api.post('/me/nickname', { newNickname });
+  return response.data;
 };
