--- client/src/features/my/hooks/useRewardHistory.ts
@@ -1,18 +1,0 @@
-import { useQuery } from '@tanstack/react-query';
-import { getRewardHistory } from '../api/getRewardHistory';
-
-interface UseRewardHistoryQueryOptions {
-  pageNum?: number;
-  pageSize?: number;
-  enabled?: boolean;
-}
-
-export const useRewardHistoryQuery = ({
-  pageNum = 0,
-  pageSize = 10,
-}: UseRewardHistoryQueryOptions = {}) => {
-  return useQuery({
-    queryKey: ['rewardHistory', pageNum, pageSize],
-    queryFn: () => getRewardHistory({ pageNum, pageSize }),
-  });
-};
