+++ client/src/features/my/ui/ChangeNicknameForm.tsx
@@ -1,5 +1,4 @@
 import { useRandomNicknameQuery } from '@/features/auth/hooks/useRandomNicknameQuery';
-import { useChangeNicknameMutation } from '../hooks/useChangeNicknameMutation';
 import { useEnterKeyHandler } from '@/shared/hooks';
 import { useEffect } from 'react';
 import * as S from './ChangeNicknameForm.styles';
@@ -7,6 +6,7 @@ import { RotateNicknameButton } from '@/widgets/signup/SignupStep2';
 import { RotateCw } from 'lucide-react';
 import { Input } from '@/shared/ui/input/Input';
 import { Button } from '@/shared/ui';
+import { useChangeNicknameMutation } from '../api/useChangeNicknameMutation';
 
 export const ChangeNicknameForm = ({
   nickname,
